requirements.txt
fastmcp
python-dotenv
requests
pytest
slack-sdk

